CmdHost filter
