+++ code/system.ui/react.common/src/ui.dev/Dev.CmdHost.Stateful/common.ts
@@ -1,1 +1,18 @@
-export * from '../Dev.CmdHost/common';
+import { DEFAULTS as BASE, type t } from '../common';
+import { Filter } from '../Dev.CmdHost/common.Filter';
+
+export * from '../common';
+export { Filter };
+
+/**
+ * Constants
+ */
+const filter: t.CmdHostFilter = (imports, command) => {
+  return Filter.imports(imports, command, { maxErrors: 1 });
+};
+
+export const DEFAULTS = {
+  displayName: 'CmdHost.Stateful',
+  filter,
+  qs: BASE.qs,
+} as const;
