+++ code/system.ui/react.common/src/ui.dev/Dev.CmdHost/common.ts
@@ -1,4 +1,5 @@
 import { DEFAULTS as BASE, DevBase, type t } from '../common';
+import { Filter } from './common.Filter';
 
 export { CmdBar } from '../../ui/Cmd.Bar';
 export * from '../common';
@@ -10,13 +11,15 @@ export const ModuleList = DevBase.ModuleList;
  * Constants
  */
 const pkg: t.CmdHostProps['pkg'] = { name: 'unknown', version: '0.0.0' };
+const filter: t.CmdHostFilter = (imports, command) => Filter.imports(imports, command);
 
 export const DEFAULTS = {
+  displayName: 'CmdHost',
   pkg,
+  filter,
   commandPlaceholder: 'command',
   focusOnReady: true,
   focusOnClick: false,
   autoGrabFocus: true,
-  applyFilter: true,
   qs: BASE.qs,
 } as const;
